Build the SiWx917 SoC in the new wifi structure
