+++ src/platform/silabs/SiWx917/BUILD.gn
@@ -30,10 +30,27 @@ if (chip_crypto == "platform") {
   import("//build_overrides/mbedtls.gni")
 }
 
-config("siwx917-platform-wifi-config") {
-  include_dirs = [ "wifi" ]
-  defines = []
+source_set("logging") {
+  sources = [ "${chip_root}/src/platform/silabs/Logging.cpp" ]
+  deps = [
+    "${chip_root}/src/platform:platform_base",
+    "${chip_root}/src/platform/logging:headers",
+  ]
 }
+
+# source_set("wifi-interface") {
+
+#   public_deps = [
+#     "${chip_root}/src/app/icd/server:icd-server-config",
+#     "${chip_root}/src/lib/core",
+#     "${chip_root}/src/lib/support",
+#     "${chip_root}/src/platform:platform",
+#     "${chip_root}/src/platform/silabs/SiWx917:SiWx917",
+#     "${lwip_root}:lwip",
+#   ]
+
+# }
+
 static_library("SiWx917") {
   sources = [
     "${silabs_platform_dir}/BLEManagerImpl.h",
@@ -45,6 +62,7 @@ static_library("SiWx917") {
     "${silabs_platform_dir}/ConfigurationManagerImpl.cpp",
     "${silabs_platform_dir}/ConfigurationManagerImpl.h",
     "${silabs_platform_dir}/ConnectivityManagerImpl.h",
+    "${silabs_platform_dir}/ConnectivityManagerImpl_WIFI.cpp",
     "${silabs_platform_dir}/DiagnosticDataProviderImpl.cpp",
     "${silabs_platform_dir}/DiagnosticDataProviderImpl.h",
     "${silabs_platform_dir}/InetPlatformConfig.h",
@@ -52,6 +70,8 @@ static_library("SiWx917") {
     "${silabs_platform_dir}/KeyValueStoreManagerImpl.h",
     "${silabs_platform_dir}/MigrationManager.cpp",
     "${silabs_platform_dir}/MigrationManager.h",
+    "${silabs_platform_dir}/NetworkCommissioningWiFiDriver.cpp",
+    "${silabs_platform_dir}/NetworkCommissioningWiFiDriver.h",
     "${silabs_platform_dir}/PlatformManagerImpl.cpp",
     "${silabs_platform_dir}/PlatformManagerImpl.h",
     "${silabs_platform_dir}/SilabsConfig.cpp",
@@ -66,6 +86,7 @@ static_library("SiWx917") {
     "../rs911x/rsi_ble_config.h",
     "../rs911x/wfx_sl_ble_init.cpp",
     "../rs911x/wfx_sl_ble_init.h",
+    "SiWxPlatformInterface.h",
   ]
 
   if (chip_enable_ota_requestor) {
@@ -94,23 +115,7 @@ static_library("SiWx917") {
     public_deps += [
       "${chip_root}/src/crypto",
       "${mbedtls_root}:mbedtls",
+      "${silabs_platform_dir}/wifi:wifi-platform",
     ]
   }
-
-  sources += [
-    "${silabs_platform_dir}/ConnectivityManagerImpl_WIFI.cpp",
-    "${silabs_platform_dir}/NetworkCommissioningWiFiDriver.cpp",
-    "${silabs_platform_dir}/NetworkCommissioningWiFiDriver.h",
-    "${silabs_platform_dir}/wifi/wfx_host_events.h",
-    "wifi/wfx_msgs.h",
-  ]
-  public_configs = [ ":siwx917-platform-wifi-config" ]
-}
-
-source_set("logging") {
-  sources = [ "${chip_root}/src/platform/silabs/Logging.cpp" ]
-  deps = [
-    "${chip_root}/src/platform:platform_base",
-    "${chip_root}/src/platform/logging:headers",
-  ]
 }
