+++ src/platform/silabs/wifi/BUILD.gn
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import("//build_overrides/chip.gni")
+import("//build_overrides/lwip.gni")
 import("${chip_root}/src/platform/device.gni")
 import("${chip_root}/third_party/silabs/efr32_sdk.gni")
 import("${chip_root}/third_party/silabs/silabs_board.gni")
@@ -32,7 +33,7 @@ declare_args() {
   chip_default_wifi_psk = ""
 }
 
-if (chip_enable_wifi) {
+if (chip_enable_wifi && !wifi_soc) {
   assert(use_rs9116 || use_wf200 || use_SiWx917)
   import("${chip_root}/src/platform/silabs/wifi_args.gni")
 
@@ -61,6 +62,7 @@ config("wifi-platform-config") {
     include_dirs += rs9117_inc_plat
   } else if (use_wf200) {
     include_dirs += wf200_plat_incs
+  } else {
   }
 
   if (chip_default_wifi_ssid != "") {
@@ -107,6 +109,14 @@ source_set("wifi-platform") {
     "${silabs_platform_dir}/wifi/wfx_msgs.h",
   ]
 
+  public_configs = [ ":wifi-platform-config" ]
+
+  public_deps = [
+    "${chip_root}/src/app/icd/server:icd-server-config",
+    "${chip_root}/src/inet",
+    "${chip_root}/src/lib/support",
+  ]
+
   cflags = []
 
   if (use_rs9116) {
@@ -128,6 +138,19 @@ source_set("wifi-platform") {
     sources += wf200_plat_src
   }
 
+  if (wifi_soc) {
+    sources += [
+      "${silabs_platform_dir}/wifi/SiWx/SiWxWifiInterface.cpp",
+      "${silabs_platform_dir}/wifi/wiseconnect-abstraction/WiseconnectInterfaceAbstraction.cpp",
+      "${silabs_platform_dir}/wifi/wiseconnect-abstraction/WiseconnectInterfaceAbstraction.h",
+
+      # Wi-Fi Config - Using the file sdk support until the wiseconnect file is fixed
+      "${sdk_support_root}/components/service/network_manager/src/sl_net_for_lwip.c",
+    ]
+
+    public_deps += [ "${lwip_root}:lwip" ]
+  }
+
   if (use_wf200 || use_rs9116) {
     sources += [
       "${silabs_platform_dir}/wifi/lwip-support/dhcp_client.cpp",
@@ -136,12 +159,4 @@ source_set("wifi-platform") {
       "${silabs_platform_dir}/wifi/lwip-support/lwip_netif.cpp",
     ]
   }
-
-  public_configs = [ ":wifi-platform-config" ]
-
-  public_deps = [
-    "${chip_root}/src/app/icd/server:icd-server-config",
-    "${chip_root}/src/inet",
-    "${chip_root}/src/lib/support",
-  ]
 }
